Remove `::bigint` cast for long literal formatting
The `::bigint` casts were confusing if dealing with timestamp or date
`Long` values.

And it isn't required. Numeric literals parsed in the `Integer` range
are implicitly cast to `bigint` if needed.

+++ server/src/test/java/io/crate/planner/operators/LimitTest.java
@@ -73,8 +73,8 @@ public class LimitTest extends CrateDummyClusterServiceUnitTest {
         );
         assertThat(plan).isEqualTo(
             """
-            Limit[20::bigint;7::bigint]
-              └ Limit[10::bigint;5::bigint]
+            Limit[20;7]
+              └ Limit[10;5]
                 └ Collect[doc.users | [name] | true]
             """
         );
